Add webpack file-loader to copy index.html
However, it isn't really a task. You need to import/require the
necessary file, so it is part of the dependency graph, and the loader is
configured to copy out to the output path ("dist") in webpack.config.js.

Without the file-loader module installed, the index.html file will be
processed as a javascript file, and would cause issues with the build
process.