Vendor compiled angular_1_router.js (ngComponentRouter)
Compiled with checkout at git commit 4087e31:
https://github.com/angular/angular/commit/4087e31

Mostly to get the huge benefit added in this commit a few days ago
https://github.com/angular/angular/commit/fde026a9e485d9e68b408793add7eaffcf383a86

Allowing you to refer to a directive component by name instead of controller class.

This makes it wayyy more like Angular 2.

Caveat: You must declare your component directives with `restrict: ‘A’` (or some variation that includes ‘A’) if you want to route to them with ngComponentRouter, since it replaces the `component` key in a route object like:

{ path: '/about', component: 'about' }

with `<div about></div>`

instead of `<about></about>`, so `restrict ‘E’` directives won’t match.